ds: refactor types

+++ src/libtrader/ds/asset/stock.rs
@@ -2,23 +2,23 @@ use chrono::{DateTime, Utc};
 
 #[derive(PartialEq, Debug)]
 pub struct StockVal {
-    pub open_val: i64,
-    pub close_val: i64,
-    pub low_val: i64,
-    pub high_val: i64,
-    pub avg_val: i64,
-    pub vol: i64,
+    pub open_val: f64,
+    pub close_val: f64,
+    pub low_val: f64,
+    pub high_val: f64,
+    pub avg_val: f64,
+    pub vol: isize,
     pub open_date: DateTime<Utc>,
     pub close_date: DateTime<Utc>,
 }
 
 #[derive(PartialEq, Debug)]
 pub struct StockDayVal {
-    pub open_val: i64,
-    pub close_val: i64,
-    pub low_val: i64,
-    pub high_val: i64,
-    pub vol: i64,
+    pub open_val: f64,
+    pub close_val: f64,
+    pub low_val: f64,
+    pub high_val: f64,
+    pub vol: isize,
     pub open_date: DateTime<Utc>,
     pub close_date: DateTime<Utc>,
     pub stock_vals: Vec<StockVal>,
@@ -26,11 +26,11 @@ pub struct StockDayVal {
 
 #[derive(PartialEq, Debug)]
 pub struct StockWeekVal {
-    pub open_val: i64,
-    pub close_val: i64,
-    pub low_val: i64,
-    pub high_val: i64,
-    pub vol: i64,
+    pub open_val: f64,
+    pub close_val: f64,
+    pub low_val: f64,
+    pub high_val: f64,
+    pub vol: isize,
     pub open_date: DateTime<Utc>,
     pub close_date: DateTime<Utc>,
     pub stock_vals: [StockDayVal; 5],
@@ -38,11 +38,11 @@ pub struct StockWeekVal {
 
 #[derive(PartialEq, Debug)]
 pub struct StockMonthVal {
-    pub open_val: i64,
-    pub close_val: i64,
-    pub low_val: i64,
-    pub high_val: i64,
-    pub vol: i64,
+    pub open_val: f64,
+    pub close_val: f64,
+    pub low_val: f64,
+    pub high_val: f64,
+    pub vol: isize,
     pub open_date: DateTime<Utc>,
     pub close_date: DateTime<Utc>,
     pub week_vals: [StockWeekVal; 4],
@@ -50,11 +50,11 @@ pub struct StockMonthVal {
 
 #[derive(PartialEq, Debug)]
 pub struct StockYearVal {
-    pub open_val: i64,
-    pub close_val: i64,
-    pub low_val: i64,
-    pub high_val: i64,
-    pub vol: i64,
+    pub open_val: f64,
+    pub close_val: f64,
+    pub low_val: f64,
+    pub high_val: f64,
+    pub vol: isize,
     pub open_date: DateTime<Utc>,
     pub close_date: DateTime<Utc>,
     pub month_vals: [StockMonthVal; 12],
@@ -63,6 +63,6 @@ pub struct StockYearVal {
 #[derive(Default, PartialEq, Debug)]
 pub struct Stock {
     pub symbol: String,
-    pub latest_price: i64,
+    pub latest_price: f64,
     pub val: Vec<StockYearVal>
 }
